Decompose sidebar file
